% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{cdm.mu.bcd}
\alias{cdm.mu.bcd}
\title{given the sample of a single variable the centered distance matrix, mu and bcd are computed
The normalization should be posponed to the moment calulation.}
\usage{
cdm.mu.bcd(x, normalize = FALSE, psi = NULL, p = NULL,
  isotropic = FALSE, unbiased.moments = TRUE)
}
\description{
NOTE: speedup might be possible by incorporating mu3 and some matrix-norm-identities
}
\keyword{internal}
