% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{cdms.mu.bcd}
\alias{cdms.mu.bcd}
\title{computes the centered distance matrices, mus and bcds}
\usage{
cdms.mu.bcd(x, vec = 1:ncol(x), membership = NULL, ...)
}
\arguments{
\item{x}{matrix, each row is a sample}

\item{vec}{vector which indicates which columns are treated as one sample}

\item{membership}{depreciated. Now use \code{vec}.}

\item{...}{these are passed to \code{\link{cdm}}}
}
\value{
It returns an 3 dimensional array of the distance matrices. The index of the first dimension names the component for which the matrix is computed, thus it ranges from 1 to max(vec).
}
\description{
computes the centered distance matrices, mus and bcds
}
\keyword{internal}
