% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{cdms}
\alias{cdms}
\title{computes the centered distance matrices}
\usage{
cdms(x, membership = 1:ncol(x), ...)
}
\arguments{
\item{x}{matrix, each row is a sample}

\item{membership}{vector which indicates which columns are treated as one sample}

\item{...}{these are passed to \code{\link{cdm}}}
}
\value{
It returns an 3 dimensional array of the distance matrices. The index of the first dimension names the component for which the matrix is computed, thus it ranges from 1 to max(membership).
}
\description{
computes the centered distance matrices
}
\details{
\code{membership} has to be a vector of length \code{ncol(x)}. Furthermore it has to have values from 1 to the number of classes.
}
