% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{circle.coordinates}
\alias{circle.coordinates}
\title{calculates the coordinates of n points on a circle of radius r
if only 1 inner point, then it is placed in the center}
\usage{
circle.coordinates(n, r = 0.5, add.angle = 0)
}
\description{
calculates the coordinates of n points on a circle of radius r
if only 1 inner point, then it is placed in the center
}
\keyword{internal}
