% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{clean.graph}
\alias{clean.graph}
\title{cleanup dependence structure graph}
\usage{
clean.graph(g)
}
\arguments{
\item{g}{graph, created by \code{\link{dependence.structure}}}
}
\value{
graph
}
\description{
Given a dependence structure graph: vertices representing the multivariances of only two vertices become an edge labeled with the label of the vertex.
}
\examples{

N = 200
y = coins(N,2)
x = cbind(y,y,y)
ds = dependence.structure(x)
plot(clean.graph(ds$graph))
}
