% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{coins}
\alias{coins}
\title{k-independent coin sampling}
\usage{
coins(N = 1000, k = 2)
}
\arguments{
\item{N}{number of samples}

\item{k}{each k-tuple will be independent}
}
\value{
It returns the samples as rows of an \code{N} by \code{k+1} matrix. The columns are dependent but k-independent.
}
\description{
This function creates samples which are k-independent.
}
\examples{
coins(200,4)

}
