% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{copula.multicorrelation.test}
\alias{copula.multicorrelation.test}
\title{independence tests using the copula versions of distance multivariance}
\usage{
copula.multicorrelation.test(x, vec = 1:ncol(x), ...)
}
\arguments{
\item{x}{matrix, each row is a sample}

\item{vec}{vector which indicates which columns are treated as one sample}

\item{...}{these are passed to \code{\link{cdm}}}
}
\description{
Formally it is nothing but tests for distance multivariance applied to the Monte Carlo emprical transform of the data. Hence its values vary for repeated runs.
}
\references{
For the theoretic background see the reference [5] given on the main help page of this package: \link{multivariance-package}.
}
