% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\docType{data}
\name{dep_struct_ring_15_100}
\alias{dep_struct_ring_15_100}
\title{example dataset for \code{\link{dependence.structure}}}
\format{
\code{matrix} 15 variables (columns), 100 independent samples (rows)
}
\usage{
dep_struct_ring_15_100
}
\description{
It was generated by \preformatted{
set.seed(436646700)
N = 100
n= 15
x=matrix(sample.int(2,N*n,replace = TRUE)-1,nrow =N)
x[,4] = rowSums(x[,1:3]) %% 2
x[,7] = rowSums(x[,4:6]) %% 2
x[,10] = rowSums(x[,7:9]) %% 2
x[,13] = rowSums(x[,10:12]) %% 2
x[,15] = rowSums(x[,c(13,14,1)]) %% 2
dep_struct_ring_15_100 = x
save(dep_struct_ring_15_100,file ="dep_struct_ring_15_100.rda")
}
}
\details{
To avoid irritation, note that the seed is just a simple integer hash value of the variable name.
}
\keyword{datasets}
