% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\docType{data}
\name{dep_struct_several_26_100}
\alias{dep_struct_several_26_100}
\title{example dataset for \code{\link{dependence.structure}}}
\format{
\code{matrix} 26 variables (columns), 100 independent samples (rows)
}
\usage{
dep_struct_several_26_100
}
\description{
It was generated by \preformatted{
set.seed(1348879148)
N = 100
dep_struct_several_26_100 = cbind(coins(N,2),tetrahedron(N),coins(N,4),
    tetrahedron(N),tetrahedron(N),coins(N,3),coins(N,3),rnorm(N))
save(dep_struct_several_26_100,file ="dep_struct_several_26_100.rda")
}
}
\details{
To avoid irritation, note that the seed is just a simple integer hash value of the variable name.
}
\keyword{datasets}
