% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\docType{data}
\name{dep_struct_star_9_100}
\alias{dep_struct_star_9_100}
\title{example dataset for \code{\link{dependence.structure}}}
\format{
\code{matrix} 9 variables (columns), 100 independent samples (rows)
}
\usage{
dep_struct_star_9_100
}
\description{
It was generated by \preformatted{
set.seed(222454572)
N = 100
y = coins(N,2)
dep_struct_star_9_100 = cbind(y,y,y)
save(dep_struct_star_9_100,file ="dep_struct_star_9_100.rda")
}
}
\details{
To avoid irritation, note that the seed is just a simple integer hash value of the variable name.
}
\keyword{datasets}
