% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{dependence.structure.full}
\alias{dependence.structure.full}
\title{functions to detect the full (without clustering) dependence structure}
\usage{
dependence.structure.full(
  x,
  vec = 1:ncol(x),
  verbose = TRUE,
  type = "conservative",
  alpha = 0.05,
  list.cdm = NULL,
  maxk = max(vec),
  stop.too.many = NULL,
  c.factor = 2,
  ...
)
}
\description{
functions to detect the full (without clustering) dependence structure
}
\examples{
# multivariance:::dependence.structure.full(dep_struct_ring_15_100)
# dependence.structure(dep_struct_ring_15_100,structure.type = "full")

}
\keyword{internal}
