% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{dist.to.matrix}
\alias{dist.to.matrix}
\title{transforms a distance matrix to a matrix}
\usage{
dist.to.matrix(ds)
}
\arguments{
\item{ds}{a distance matrix object, e.g. generated by \code{\link{dist}}}
}
\description{
Does for a distance matrix generated via \code{dist} the same as \code{as.matrix} only slightly faster.
}
\keyword{internal}
