% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{dm}
\alias{dm}
\title{distance matrix}
\usage{
dm(x, psi = NULL, p = NULL, isotropic = FALSE, external.dm.fun = NULL)
}
\description{
# currently only used for the bias corrected multicorrelations
# it might be used globally to remove redundancies
}
\keyword{internal}
