% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{dms}
\alias{dms}
\title{list of distance matrices}
\usage{
dms(x, vec = 1:ncol(x), ...)
}
\description{
# currently only used for the bias corrected multicorrelations
# it might be used globally, to remove redundancies
}
\keyword{internal}
