% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{double.center}
\alias{double.center}
\title{double centering of a matrix}
\usage{
double.center(dm, normalize = TRUE)
}
\description{
# changed default after 2.0.0
}
\keyword{internal}
