% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doubleCenterBiasCorrected}
\alias{doubleCenterBiasCorrected}
\title{bias corrected double centering
# included for speed comparison}
\usage{
doubleCenterBiasCorrected(x)
}
\description{
bias corrected double centering
# included for speed comparison
}
\keyword{internal}
