% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doubleCenterBiasCorrectedUpper}
\alias{doubleCenterBiasCorrectedUpper}
\title{bias corrected double centering with normalizing
# included for speed comparison}
\usage{
doubleCenterBiasCorrectedUpper(x)
}
\description{
bias corrected double centering with normalizing
# included for speed comparison
}
\keyword{internal}
