% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doubleCenterBiasCorrectedUpperLower}
\alias{doubleCenterBiasCorrectedUpperLower}
\title{bias corrected double centering with normalizing constants for upper and lower bound}
\usage{
doubleCenterBiasCorrectedUpperLower(x, n)
}
\description{
bias corrected double centering with normalizing constants for upper and lower bound
}
\keyword{internal}
