% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doubleCenterSymMat}
\alias{doubleCenterSymMat}
\title{double center a symmetric matrix}
\usage{
doubleCenterSymMat(x, normalize)
}
\arguments{
\item{x}{symmetric matrix}

\item{normalize}{boolean. If \code{TRUE} the matrix will be normalized to mean 1.}
}
\description{
double center a symmetric matrix
}
\keyword{internal}
