% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{emp.transf.dep}
\alias{emp.transf.dep}
\title{A dependent Monte Carlo emprical transform}
\usage{
emp.transf.dep(x)
}
\description{
A dependent Monte Carlo emprical transform
}
\keyword{internal}
