% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{emp.transf}
\alias{emp.transf}
\title{Monte Carlo empirical transform}
\usage{
emp.transf(x, continuous = FALSE)
}
\arguments{
\item{x}{data matrix (rows: samples, columns: variables)}

\item{continuous}{boolean, if TRUE it provides the classical (non-Monte-Carlo) transformation by the empirical distribution function, which is a reasonable choice for data of continuous distributions.}
}
\description{
Transforms a matrix (rows: samples, columns: variables) into a matrix of uniform samples with the same dependence structure via the Monte Carlo empirical transform.
}
\references{
For the theoretic background see the reference [5] given on the main help page of this package: \link{multivariance-package}.
}
