% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{emp.transf.vec}
\alias{emp.transf.vec}
\title{Transform a vector of samples into a vector of samples of the uniform distribution
such that, if applied to multiple (dependent) sample vectors, the dependence is preserved.}
\usage{
emp.transf.vec(x, unif.samples = stats::runif(length(x)))
}
\description{
Transform a vector of samples into a vector of samples of the uniform distribution
such that, if applied to multiple (dependent) sample vectors, the dependence is preserved.
}
\examples{
# x = rnorm(100)
# plot(emp.transf.vec(x))

}
\keyword{internal}
