% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastdist}
\alias{fastdist}
\title{fast euclidean distance matrix computation}
\usage{
fastdist(x)
}
\arguments{
\item{x}{the vector for which the distanc matrix is computed}
}
\description{
fast euclidean distance matrix computation
}
\examples{
#require(microbenchmark)
#x = rnorm(100)
#microbenchmark(fastdist(as.matrix(x)),as.matrix(dist(x)))
}
