% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{is.doubly.centered}
\alias{is.doubly.centered}
\title{checks if a matrix is doubly centered}
\usage{
is.doubly.centered(mat)
}
\arguments{
\item{mat}{matrix}
}
\description{
!works only for the biased estimators
}
\examples{
multivariance:::is.doubly.centered(as.matrix(dist(rnorm(10))))
multivariance:::is.doubly.centered(multivariance:::double.center(as.matrix(dist(rnorm(10)))))

}
\keyword{internal}
