% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{lower.order}
\alias{lower.order}
\title{check if lower order dependencies are present for the given tuple indices
here 'm.values' is a list of boolean matrices. Matrix [[k]] corresponds to the k tuples. For each number of tuples, the first columns of the matrix always contain the indices of the tuples}
\usage{
lower.order(tuple, m.values)
}
\description{
check if lower order dependencies are present for the given tuple indices
here 'm.values' is a list of boolean matrices. Matrix [[k]] corresponds to the k tuples. For each number of tuples, the first columns of the matrix always contain the indices of the tuples
}
\keyword{internal}
