% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{match.rows}
\alias{match.rows}
\title{Returns the row indices of matrix A which match with B
Use the fast cpp implementation 'match_rows' instead.
Function here just for reference.}
\usage{
match.rows(A, B)
}
\description{
Returns the row indices of matrix A which match with B
Use the fast cpp implementation 'match_rows' instead.
Function here just for reference.
}
\examples{
# A = t(utils::combn(10,3))
# B = A[sort(sample.int(nrow(A),10)),]
# match.rows(A,B)

}
\keyword{internal}
