% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_rows}
\alias{match_rows}
\title{for the fast detection of the full dependence structure}
\usage{
match_rows(A, B)
}
\arguments{
\item{A}{matrix}

\item{B}{matrix whose rows are subset of A}
}
\description{
Returns the row indicies of matrix A which match with B
}
\examples{
# A = t(utils::combn(10,3))
# B = A[sort(sample.int(nrow(A),10)),]
# match_rows(A,B)

}
\keyword{internal}
