% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{moments.for.pearson}
\alias{moments.for.pearson}
\title{computes the moments as required by pearson}
\usage{
moments.for.pearson(N, bcd, mu, mmean, type = "multi")
}
\description{
computes the moments as required by pearson
}
\keyword{internal}
