% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{mu3.unbiased}
\alias{mu3.unbiased}
\title{given the distance matrix the unbiased estimate for mu3 is computed}
\usage{
mu3.unbiased(B, b2ob = sum(tcrossprod(B) * B))
}
\description{
given the distance matrix the unbiased estimate for mu3 is computed
}
\keyword{internal}
