% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{multicorrelation.bias.corrected}
\alias{multicorrelation.bias.corrected}
\title{bias corrected total multicorrelations}
\usage{
multicorrelation.bias.corrected(
  x,
  vec = 1:ncol(x),
  squared = FALSE,
  type = "all",
  ...
)
}
\description{
bias corrected total multicorrelations
}
\keyword{internal}
