% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{multivariance.pvalue}
\alias{multivariance.pvalue}
\title{transform multivariance to pvalue}
\usage{
multivariance.pvalue(x)
}
\arguments{
\item{x}{value of a normalized and Nscaled \code{\link{multivariance}}}
}
\description{
Computes the p-value (under the hypothesis of independence) for a given multivariance/total multivariance.
}
\details{
The p-value is conservative, i.e. it might be much smaller. This is the counterpart to \code{\link{rejection.level}}.
}
