% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{N.coefficients}
\alias{N.coefficients}
\title{Computes the explicit coefficients for the finite sample variance for a sample of size N}
\usage{
N.coefficients(N)
}
\description{
Computes the explicit coefficients for the finite sample variance for a sample of size N
}
\keyword{internal}
