% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{p.value.to.star.label}
\alias{p.value.to.star.label}
\title{transforms a p-value into the corresponding label}
\usage{
p.value.to.star.label(pv)
}
\arguments{
\item{pv}{p-value}
}
\description{
transforms a p-value into the corresponding label
}
\keyword{internal}
