% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{pairwise.multicorrelation.bias.corrected}
\alias{pairwise.multicorrelation.bias.corrected}
\title{pairwise multicorrelation}
\usage{
pairwise.multicorrelation.bias.corrected(
  x,
  vec = 1:ncol(x),
  squared = FALSE,
  ...
)
}
\description{
pairwise multicorrelation
}
\keyword{internal}
