% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{pearson.qf}
\alias{pearson.qf}
\title{approximate distribution function of a Gaussian quadratic form}
\usage{
pearson.qf(x, moment, lower.tail = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{value at which the distribution function is to be evaluated}

\item{moment}{vector with the mean, variance and skewness of the quadratic form}

\item{lower.tail}{logical, indicating of the lower or upper tail of the distribution function should be calculated}

\item{verbose}{logical, if \code{TRUE} a warning is issued if negative moments are sanitized to 0.}
}
\description{
Approximation of the of the value of the distribution function of a Gaussian quadratic form based on its first three moments.
}
\details{
This is Pearson's approximation for Gaussian quadratic forms as stated in [4] (equation (4.65) in arXiv:1808.07280v2)
}
\references{
For the theoretic background see the reference [4] given on the main help page of this package: \link{multivariance-package}.
}
