% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{resample.rejection.level}
\alias{resample.rejection.level}
\title{rejection level via resampling}
\usage{
resample.rejection.level(alpha = 0.05, ...)
}
\arguments{
\item{alpha}{numeric, the significance value}

\item{...}{passed to \code{\link{resample.multivariance}}. Required is the data matrix \code{x}.}
}
\description{
Uses the resample method to sample from the test statistic under the hypothesis of independence. The alpha quantile of these samples is returned.
}
\examples{
resample.rejection.level(0.05,matrix(rnorm(30*2),nrow = 30))
resample.rejection.level(0.05,matrix(rnorm(30*3),nrow = 30),vec = c(1,1,2))

}
\references{
For the theoretic background see the reference [3] given on the main help page of this package: \link{multivariance-package}.
}
