% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{sample.cdms}
\alias{sample.cdms}
\title{resamples centered distance matricies}
\usage{
sample.cdms(array.cdm)
}
\arguments{
\item{array.cdm}{an array of centered distance matricies}
}
\value{
Returns an array of centered distance matricies, each matrix corresponds to the resampled columns of the corresponding sample, using resampling without replacement (permutations).
}
\description{
resamples centered distance matricies
}
