% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{simple.int.hash}
\alias{simple.int.hash}
\title{Simple integer hash from text}
\usage{
simple.int.hash(x)
}
\description{
Used to compute a random seed for \code{set.seed} based on the example name,
in order to aviod arbitrary seeds like '1234'.
}
\examples{
multivariance:::simple.int.hash("dep_struct_several_26_100")
}
\keyword{internal}
