% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{sums.of.products}
\alias{sums.of.products}
\title{This is the function GC which is required for the computation of the finite sample variance for m and total multivariance}
\usage{
sums.of.products(a, b, c, type = "multi")
}
\description{
This is the function GC which is required for the computation of the finite sample variance for m and total multivariance
}
\keyword{internal}
