% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{total.multicorrelation.bias.corrected.upper}
\alias{total.multicorrelation.bias.corrected.upper}
\title{# included for speed. it is faster than upper.lower}
\usage{
total.multicorrelation.bias.corrected.upper(
  x,
  vec = 1:ncol(x),
  squared = FALSE,
  ...
)
}
\description{
# included for speed. it is faster than upper.lower
}
\keyword{internal}
