% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{cdm.mu.bcd}
\alias{cdm.mu.bcd}
\title{given the sample of a single variable the doubly centered distance matrix, mu (the limit moments) and bcd (the terms for the finite sample moments) are computed}
\usage{
cdm.mu.bcd(
  x,
  normalize = FALSE,
  psi = NULL,
  p = NULL,
  isotropic = FALSE,
  unbiased.moments = TRUE,
  external.dm.fun = NULL
)
}
\description{
The normalization should be postponed to the moment calculation.
}
\keyword{internal}
