% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{cdms}
\alias{cdms}
\title{computes the doubly centered distance matrices}
\usage{
cdms(x, vec = 1:ncol(x), membership = NULL, ...)
}
\arguments{
\item{x}{matrix, each row is a sample}

\item{vec}{vector which indicates which columns are treated as one sample}

\item{membership}{depreciated. Now use \code{vec}.}

\item{...}{these are passed to \code{\link{cdm}}}
}
\value{
It returns a list of distance matrices.
}
\description{
computes the doubly centered distance matrices
}
