% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{moments.for.pearson}
\alias{moments.for.pearson}
\title{computes the moments as required for Pearson's approximation}
\usage{
moments.for.pearson(N, bcd, mu, mmean, type = "multi")
}
\arguments{
\item{N}{sample size}

\item{bcd}{an array with b c d}

\item{mu}{the limit moments}

\item{mmean}{the means of the distance matrices}
}
\description{
computes the moments as required for Pearson's approximation
}
\details{
Note: It is currently only implemented for the case of normalized multivariance, i.e., the OUTPUT values correspond to normalized multivariance!
}
\keyword{internal}
