% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariance-functions.R
\name{sample.cdms}
\alias{sample.cdms}
\title{resamples centered distance matrices}
\usage{
sample.cdms(list.cdm, replace = FALSE)
}
\arguments{
\item{list.cdm}{a list of centered distance matrices}

\item{replace}{boolean, sampling with or without replacement}
}
\value{
Returns a list of centered distance matrices, each matrix corresponds to the resampled columns of the corresponding sample, using resampling with replacement (bootstrap) or without replacement (permutations).
}
\description{
resamples centered distance matrices
}
